requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_jwt_extended
flask_cors
python-dotenv
bcrypt
pymysql